discordBot/requirements.txt
discord.py[voice]
python-dotenv
sqlalchemy
python-dotenv
psycopg2
langdetect
youtube-dl
emoji